Add support for the 'url' type
This allows you to add a new mapping the the mapping dictionary
that is of type 'url'. This will drop in a reference directive
this is particularly useful for mapping a key to an external
set of URLS. For example, if you frequently need to reference
an external project, you can map directly to their types
using the mapping.